3단계 코드리뷰 반영
